fixed 404 on blogs
